Prompt 4: Anchor protocol config, observer sets and market registry
- program: initialize_protocol (upgrade-authority gated), two-step admin transfer,
  new-risk pause, immutable sequential ObserverSet, MarketConfig registry
- market binding reads the DLMM LbPair (mints, orientation) fail-closed, with layout
  verified against real mainnet bytes; off-chain review trust boundary documented
- mandate-core: protocol/observer validation and the pure LbPair reader
- program-tests: LiteSVM executing the compiled SBF binary, 55 tests (signer and
  account-substitution attacks, bounds, pause, admin transfer, immutability); five
  deliberate program mutations were all caught
- host Rust to 1.98.1 for LiteSVM; CI builds program + IDL before Rust tests
- ADR 0011

+++ crates/mandate-core/src/validation.rs
@@ -41,7 +41,7 @@ pub struct CreateMandateParams {
 
 /// A set of distinct error codes.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub struct ErrorSet(u16);
+pub struct ErrorSet(u32);
 
 impl ErrorSet {
     pub fn insert(&mut self, error: MandateCoreError) {
@@ -56,8 +56,8 @@ impl ErrorSet {
         self.0 == 0
     }
 
-    const fn bit(error: MandateCoreError) -> u16 {
-        match 1u16.checked_shl(error as u32) {
+    const fn bit(error: MandateCoreError) -> u32 {
+        match 1u32.checked_shl(error as u32) {
             Some(bit) => bit,
             None => 0,
         }
